Incomplete working background updates check and install

+++ app/src/main/java/dev/re7gog/b_sideloader/ui/features/app_details/AppDetailsUiState.kt
@@ -11,7 +11,7 @@ data class AppDetailsUiState(
     val stars: Int,
     val owner: String,
     val iconUrl: String,
-    val installProgress: Float? = null,
+    val installProgress: Int? = null,
     val isInstalling: Boolean = false
 )
